requirements.txt
Flask==0.11.1
Flask-MySQLdb==0.2.0
matplotlib==1.5.3
mysqlclient==1.3.9
numpy==1.11.2
pandas==0.19.1
scikit-learn==0.18.1
scipy==0.18.1
sklearn==0.0
virtualenv==15.1.0
